requirements.txt
Flask
face_recognition
numpy
pybase64
opencv-python
pillow
cmake